Move core handlers to common package
